[IonutCava] - Re-enable sorting by materials - Add some GLSL hacks to get rendering working properly on AMD - Prevent redundant vertex buffer binds
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@538 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 9fb2adc845a6c2475c1b1e89777acc3cf7655163 [formerly b32ef9fe241364a06102bc18a754c26144623278] [formerly 2b2d819072a5f4c8ae31ac2fec6cb93c24612f56 [formerly 54bc8701880a582cb282e9d0db515e6b1ad504f6]]
Former-commit-id: 6af3f51b980e0a1826e12431e95e741ecd464550 [formerly ce5d293b652cadb5e5dcb84dcdf30846e728e749]
Former-commit-id: 8237f2e44b9f13c1fc083046028bedbadde8dfa6

+++ assets/shaders/GLSL/fbPreview.glsl
@@ -49,7 +49,7 @@ void main()
     float f = dvd_ZPlanesCombined.y * 0.5;
 #endif
 
-    float depth = textureLod(texDiffuse0, _texCoord, lodLevel).r;
+    float depth = textureLod(texDiffuse0, _texCoord, uint(lodLevel)).r;
     float linearDepth = (2 * n) / (f + n - (depth) * (f - n));
     _colorOut = vec4(vec3(linearDepth), 1.0);
 }
